Add files via upload
created a simple app.py that runs a flask application, creates the models in a new local SQL. another simple script, client.py sends data to the server, validation is in place. Kept the program as simple as possible in one file for ease of reading instead of dividing into the MVC logic. Thanks for the opportunity. Fernando